<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- The `BacktestingBrokerage` will also verify there are no more pending
orders when deciding if it needs to continue scanning.
   - The issue was being caused due to the re entrant nature of the `BacktestingBrokerage.Scan()` method. `stillNeedsScan` could be `false` for the last entry of `Scan()` but deeper stack calls had added a new order.

Performance tests C# RegressionAlgorithm 5.4K trades
`PR`
9.61 seconds at 192k data points per second. Processing total of 1,842,728 data points.
9.60 seconds at 192k data points per second. Processing total of 1,842,728 data points.
9.62 seconds at 192k data points per second. Processing total of 1,842,728 data points.
`master`
9.67 seconds at 190k data points per second. Processing total of 1,842,728 data points.
9.43 seconds at 195k data points per second. Processing total of 1,842,728 data points.
9.62 seconds at 192k data points per second. Processing total of 1,842,728 data points.
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3239
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- StopLoss order being placed in OnOrderEvent wasn't being filled
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->